test(AtomHttpApi): consolidate stream coverage
