+++ packages/effect/test/reactivity/AtomHttpApi.test.ts
@@ -1,8 +1,9 @@
 import { assert, describe, it } from "@effect/vitest"
-import { Effect, Exit, Layer, Ref, Schema } from "effect"
+import { Duration, Effect, Exit, Layer, Ref, Schema, Stream } from "effect"
+import { Sse } from "effect/unstable/encoding"
 import { HttpClient, HttpClientResponse } from "effect/unstable/http"
 import type * as HttpClientError from "effect/unstable/http/HttpClientError"
-import { HttpApi, HttpApiEndpoint, HttpApiGroup } from "effect/unstable/httpapi"
+import { HttpApi, HttpApiEndpoint, HttpApiGroup, HttpApiSchema } from "effect/unstable/httpapi"
 import { Atom, AtomHttpApi, AtomRegistry, Hydration } from "effect/unstable/reactivity"
 
 const Api = HttpApi.make("api").add(
@@ -22,6 +23,12 @@ const TopLevelApi = HttpApi.make("top-level-api").add(
   HttpApiGroup.make("group", { topLevel: true }).add(
     HttpApiEndpoint.get("query", "/query"),
     HttpApiEndpoint.post("mutation", "/mutation")
+  )
+)
+
+const StreamApi = HttpApi.make("StreamApi").add(
+  HttpApiGroup.make("events").add(
+    HttpApiEndpoint.get("watch", "/watch", { success: HttpApiSchema.StreamSse({ data: Schema.String }) })
   )
 )
 
@@ -161,4 +168,56 @@ describe("AtomHttpApi", () => {
 
       unmount()
     }))
+
+  describe("real generated SSE client", () => {
+    for (const operation of ["query", "mutation"] as const) {
+      for (const [directive, body] of [["retry", "retry: 1000\n\n"], ["data", "data: \"hello\"\n\n"]] as const) {
+        it.effect(`${operation} returns a stream preserving valid ${directive}`, () =>
+          Effect.gen(function*() {
+            const requests: Array<string> = []
+            const Client = AtomHttpApi.Service()("StreamClient", {
+              api: StreamApi,
+              baseUrl: "https://example.test",
+              httpClient: Layer.succeed(
+                HttpClient.HttpClient,
+                HttpClient.make((request) => {
+                  requests.push(request.url)
+                  return Effect.succeed(HttpClientResponse.fromWeb(
+                    request,
+                    new Response(body, {
+                      headers: { "content-type": "text/event-stream" }
+                    })
+                  ))
+                })
+              )
+            })
+            const registry = yield* Effect.acquireRelease(
+              Effect.sync(() => AtomRegistry.make()),
+              (registry) => Effect.sync(() => registry.dispose())
+            )
+            const atom = operation === "query"
+              ? Client.query("events", "watch", {})
+              : Client.mutation("events", "watch")
+            yield* AtomRegistry.mount(registry, atom)
+            if (operation === "mutation") {
+              registry.set(Client.mutation("events", "watch"), {})
+            }
+            const stream = yield* AtomRegistry.getResult(registry, atom)
+            assert(Stream.isStream(stream))
+            const result = yield* Effect.exit(Stream.runCollect(stream))
+            if (directive === "retry") {
+              assert(Exit.isFailure(result))
+              assert.deepStrictEqual(
+                result.cause.reasons.map((reason) => reason._tag === "Fail" ? reason.error : reason),
+                [new Sse.Retry({ duration: Duration.millis(1000), lastEventId: undefined })]
+              )
+            } else {
+              assert(Exit.isSuccess(result))
+              assert.deepStrictEqual(result.value, ["hello"])
+            }
+            assert.deepStrictEqual(requests, ["https://example.test/watch"])
+          }))
+      }
+    }
+  })
 })
